Move FontFinder to lib
