+++ source/llassetgen/CMakeLists.txt
@@ -4,6 +4,9 @@
 
 find_package(Freetype)
 find_package(PNG)
+if(UNIX)
+    find_package(fontconfig)
+endif()
 
 
 #
@@ -42,7 +45,8 @@ set(headers
     ${include_path}/Atlas.h
     ${include_path}/Image.h
     ${include_path}/DistanceTransform.h
-	${include_path}/FntWriter.h
+    ${include_path}/FntWriter.h
+    ${include_path}/FontFinder.h
     ${include_path}/Geometry.h
     ${include_path}/Packing.h
 )
@@ -51,7 +55,8 @@ set(sources
     ${source_path}/llassetgen.cpp
     ${source_path}/Image.cpp
     ${source_path}/DistanceTransform.cpp
-	${source_path}/FntWriter.cpp
+    ${source_path}/FntWriter.cpp
+    ${source_path}/FontFinder.cpp
     ${source_path}/packing/internal/Common.cpp
     ${source_path}/packing/internal/MaxRectsPacker.cpp
     ${source_path}/packing/internal/ShelfPacker.cpp
@@ -162,6 +167,13 @@ target_link_libraries(${target}
 
     INTERFACE
 )
+
+if (UNIX)
+    target_link_libraries(${target}
+        PRIVATE
+        fontconfig
+    )
+endif ()
 
 
 #
